ref_gl: fix crash bug on lightmaps with high-magnitude scales.
These lightmaps are generated by a special qrad3 feature that can shrink
the lightmap textures for surfaces that are completely uniform in lighting
color. Turns out qrad3 was generating values that caused integer overflows in
rare circumstances. Solution is to increase the integer size used for these
calculations.

+++ source/ref_gl/r_light.c
@@ -244,7 +244,7 @@ int RecursiveLightPoint (mnode_t *node, vec3_t start, vec3_t end)
 	cplane_t	*plane;
 	vec3_t		mid;
 	msurface_t	*surf;
-	int			s, t, ds, dt;
+	int64_t		s, t, ds, dt;
 	int			i;
 	mtexinfo_t	*tex;
 	byte		*lightmap;
